+++ app/javascript/controllers/settings_controller.js
@@ -6,17 +6,15 @@ export default class extends Controller {
   // alternative to https://maplibre.org/maplibre-gl-js/docs/API/classes/TerrainControl/
   update_terrain (event) {
     const terrain = document.querySelector('#map-terrain').checked
-    if (mapProperties.terrain !== terrain) {
-      mapProperties.terrain = terrain
-      setBackgroundMapLayer(mapProperties.base_map, true)
-      mapChannel.send_message('update_map', { terrain })
-    }
+    mapProperties.terrain = terrain
+    setBackgroundMapLayer()
+    mapChannel.send_message('update_map', { terrain })
   }
 
   update_basemap (event) {
     const layerPreviews = document.querySelectorAll('.layer-preview')
-    mapProperties.base_map = event.target.dataset.layer
-    setBackgroundMapLayer(mapProperties.base_map)
+    mapProperties.base_map = event.target.dataset.baseMap
+    setBackgroundMapLayer()
     mapChannel.send_message('update_map', { base_map: mapProperties.base_map, terrain: mapProperties.terrain })
     layerPreviews.forEach(layerPreview => { layerPreview.classList.remove('active') })
     event.target.classList.add('active')