+++ app/javascript/maplibre/styles.js
@@ -105,13 +105,16 @@ const outlineColor = featureOutlineColor
 const outlineWidth = ['+', 4, lineWidth]
 const outlineWidthActive = ['+', 3, outlineWidth]
 
-const pointColor = ['coalesce', ['get', 'user_marker-color'], ['get', 'marker-color'], featureColor]
+const pointColor = ['coalesce', ['get', 'user_marker-color'], ['get', 'marker-color'],
+  ['case',
+    ['any', ['has', 'user_marker-symbol'], ['has', 'marker-symbol']],
+    'white', featureColor]]
 export const pointSize = ['to-number', ['coalesce',
   ['get', 'user_marker-size'], ['get', 'marker-size'],
   ['case',
-    ['has', 'marker-symbol'],
-    12, 6]]]
-const pointSizeActive = ['+', 1, pointSize]
+    ['any', ['has', 'user_marker-symbol'], ['has', 'marker-symbol']],
+    16, 6]]]
+const pointSizeActive = ['+', 2, pointSize]
 const pointOutlineSize = ['to-number', ['coalesce', ['get', 'user_stroke-width'], ['get', 'stroke-width'], 2]]
 export const pointOutlineSizeActive = ['+', 1, pointOutlineSize]
 const pointOutlineColor = ['coalesce', ['get', 'user_stroke'], ['get', 'stroke'], featureOutlineColor]
@@ -122,7 +125,7 @@ const pointOpacityActive = 0.9
 // in case of icon url, we don't know the size
 // default: 1/6 = 12px (2 * default radius pointSize)
 const iconSizeFactor = ['/', pointSize, 6]
-const iconSize = ['*', 1 / 7, iconSizeFactor]
+const iconSize = ['*', 1 / 8, iconSizeFactor]
 // const iconSizeActive = ['*', 1.1, iconSize] // icon-size is not a paint property
 
 export const styles = {
@@ -234,6 +237,7 @@ export const styles = {
       ['!=', 'active', 'true']
     ],
     paint: {
+      'circle-pitch-scale': 'map', // points get bigger when camera is closer
       'circle-radius': ['case',
         ['boolean', ['feature-state', 'active'], false],
         pointSizeActive,
@@ -250,14 +254,20 @@ export const styles = {
           pointOpacityActive,
           pointOpacity
         ]],
-      'circle-stroke-color': pointOutlineColor,
+      'circle-blur': 0.1,
+      'circle-stroke-color': [
+        'case',
+        ['boolean', ['feature-state', 'active'], false],
+        featureColor,
+        pointOutlineColor
+      ],
       'circle-stroke-width': [
         'case',
         ['boolean', ['feature-state', 'active'], false],
         pointOutlineSizeActive,
         pointOutlineSize
       ],
-      'circle-stroke-opacity': 0.8
+      'circle-stroke-opacity': pointOpacity + 0.2
     }
   },
   'points-hit-layer': {
@@ -289,8 +299,8 @@ export const styles = {
           ['concat', '/emojis/noto/', ['get', 'marker-symbol'], '.png'],
           '']
       ],
-      'icon-size': iconSize,
-      'icon-overlap': 'never', // https://maplibre.org/maplibre-style-spec/layers/#icon-overlap
+      'icon-size': iconSize, // cannot scale on hover because it's not a paint property
+      'icon-overlap': 'always', // https://maplibre.org/maplibre-style-spec/layers/#icon-overlap
       'icon-ignore-placement': true // other symbols can be visible even if they collide with the icon
     },
     paint: {
@@ -316,15 +326,15 @@ export const styles = {
       'text-radial-offset': [
         'match',
         ['to-string', ['has', 'marker-symbol']],
-        'true', 1.2,
+        'true', 1.4,
         0.6
       ],
       'text-justify': 'auto',
       'text-ignore-placement': false // hide on collision
     },
     paint: {
-      'text-color': ['coalesce', ['get', 'label-color'], '#fff'],
-      'text-halo-color': ['coalesce', ['get', 'label-shadow'], '#444'],
+      'text-color': ['coalesce', ['get', 'label-color'], '#000'],
+      'text-halo-color': ['coalesce', ['get', 'label-shadow'], '#fff'],
       'text-halo-width': 1
     }
   }