+++ spec/features/map_view_spec.rb
@@ -3,12 +3,12 @@ require 'rails_helper'
 describe 'Map public view' do
   let(:map) { create(:map) }
 
-  context 'with initial map rendering' do
-    before do
-      visit map_path(map.public_id)
-      expect(page).to have_css('.maplibregl-canvas')
-    end
+  before do
+    visit map_path(map.public_id)
+    expect(page).to have_css('.map[data-loaded="true"]')
+  end
 
+  context 'with initial map rendering' do
     it 'shows map view buttons' do
       expect(page).to have_css('.maplibregl-ctrl-zoom-in')
       expect(page).to have_css('.maplibregl-ctrl-zoom-out')
@@ -24,12 +24,6 @@ describe 'Map public view' do
     let!(:polygon) { create(:feature, :polygon_middle, layer: map.layer,
       title: 'Poly Title', desc: 'Poly Desc') }
     let!(:line) { create(:feature, :line_string, layer: map.layer) }
-
-    before do
-      visit map_path(map.public_id)
-      # make sure features are loaded
-      expect(page).to have_css('.map[data-loaded="true"]')
-    end
 
     it 'shows feature details on hover' do
       hover_coord('.map', 50, 50)
@@ -48,12 +42,18 @@ describe 'Map public view' do
     end
   end
 
-  context 'with features added server side' do
-    before do
-      visit map_path(map.public_id)
-      expect(page).to have_css('.map[data-loaded="true"]')
-    end
+  context 'with features that don\'t have properties' do
+    # this polygon is in the middle of nbg (default view)
+    let!(:polygon) { create(:feature, :polygon_middle, layer: map.layer, properties: nil) }
 
+    it 'shows feature details on hover' do
+      hover_coord('.map', 50, 50)
+      expect(page).to have_css('#feature-details-modal')
+      expect(page).to have_text('No title')
+    end
+  end
+
+  context 'with features added server side' do
     # feature is created after loading the map, to make sure it's loaded via websocket
     it 'receives new features via websocket channel' do
       create(:feature, :polygon_middle, layer: map.layer, title: 'New Title')
@@ -64,12 +64,9 @@ describe 'Map public view' do
   end
 
   context 'with lost websocket' do
-    before do
-      visit map_path(map.public_id)
-      expect(page).to have_css('.map[data-loaded="true"]')
-    end
-
     it 'shows warning' do
+      # TODO: unclear why there is an update happening here, but it would hide the next status message
+      expect(page).to have_text('Map view updated')
       ActionCable.server.connections.each(&:close)
       expect(page).to have_text('Connection to server lost')
     end